Revert "Cosmetic changes, plus removing deprecations from .pyx"
This reverts commit fc1275231f58984b77bb82c35d08bc7e145e8c71.

+++ pyspike/cython/cython_add.pyx
@@ -35,7 +35,7 @@ cimport numpy as np
 
 from libc.math cimport fabs
 
-DTYPE = float
+DTYPE = np.float
 ctypedef np.float_t DTYPE_t
