+++ pyspike/cython/cython_distances.pyx
@@ -36,7 +36,7 @@ from libc.math cimport fabs
 from libc.math cimport fmax
 from libc.math cimport fmin
 
-DTYPE = float
+DTYPE = np.float
 ctypedef np.float_t DTYPE_t
