+++ pyspike/cython/cython_simulated_annealing.pyx
@@ -38,7 +38,7 @@ from libc.math cimport fmod
 from libc.stdlib cimport rand
 from libc.stdlib cimport RAND_MAX
 
-DTYPE = float
+DTYPE = np.float
 ctypedef np.float_t DTYPE_t
